backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
chromadb-client
sentence-transformers
pdfplumber
anthropic
python-dotenv
pydantic-settings
python-multipart
openai